🧩 update: 用户模块

+++ src/user/user.service.ts
@@ -1,8 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { MongoRepository } from 'typeorm';
 import { ObjectId } from 'mongodb';
+import { UserPermissionDto } from './dto/user-permission.dto';
+import { UserExcludeProjectDto } from './dto/user-exclude-project-dto';
+import { UserUpdateDto } from './dto/user-update-dto';
 
 @Injectable()
 export class UserService {
@@ -17,19 +20,65 @@ export class UserService {
   }
 
   /** 更新用户 */
-  async updateUser(user: User): Promise<User> {
-    await this.userRepository.update(user.id, user);
-    return this.userRepository.findOne(new ObjectId(user.id));
+  async updateUser(newUser: UserUpdateDto): Promise<User> {
+    console.log(newUser);
+    const { id, ...restUser } = newUser;
+    const user = await this.userRepository.findOne(new ObjectId(id));
+    if (!user) throw new HttpException('用户不存在', HttpStatus.NOT_FOUND);
+    Object.keys(restUser).forEach((key) => {
+      user[key] = restUser[key];
+    });
+    delete user.id;
+    await this.userRepository.update(id, {
+      ...user,
+      updateTime: new Date().getTime(),
+    });
+    return this.userRepository.findOne(new ObjectId(id));
   }
 
   /** 删除用户 */
-  async deleteUser(id: ObjectId): Promise<any> {
-    return this.userRepository.delete(id);
+  async deleteUser(ownerId: ObjectId, id: ObjectId): Promise<any> {
+    if (ownerId.toString() === id.toString())
+      throw new HttpException('不能删除自己', HttpStatus.BAD_REQUEST);
+    const user = await this.userRepository.findOne(new ObjectId(id));
+    if (!user) throw new HttpException('用户不存在', HttpStatus.NOT_FOUND);
+    await this.userRepository.delete(id);
+    return true;
   }
 
-  /** 根据id获取用户 */
-  async getUserById(id: ObjectId): Promise<User> {
-    const result = await this.userRepository.findOneBy(new ObjectId(id));
-    return result;
+  /** 设置用户权限 */
+  async setPermission(body: UserPermissionDto): Promise<any> {
+    const { id, type, permissions } = body;
+    const user = await this.userRepository.findOne(new ObjectId(id));
+    if (!user) throw new HttpException('用户不存在', HttpStatus.NOT_FOUND);
+    await this.userRepository.update(user.id, {
+      ...user,
+      permissions:
+        type === 1
+          ? [...new Set([...user.permissions, ...permissions])]
+          : user.permissions.filter(
+              (permission) => !permissions.includes(permission),
+            ),
+      updateTime: new Date().getTime(),
+    });
+    return this.userRepository.findOne(new ObjectId(user.id));
+  }
+
+  /** 设置不拥有的项目 */
+  async setExcludeProjects(body: UserExcludeProjectDto): Promise<any> {
+    const { id, type, excludeProjects } = body;
+    const user = await this.userRepository.findOne(new ObjectId(id));
+    if (!user) throw new HttpException('用户不存在', HttpStatus.NOT_FOUND);
+    await this.userRepository.update(user.id, {
+      ...user,
+      excludeProjects:
+        type === 1
+          ? [...new Set([...user.excludeProjects, ...excludeProjects])]
+          : user.excludeProjects.filter(
+              (excludeProject) => !excludeProjects.includes(excludeProject),
+            ),
+      updateTime: new Date().getTime(),
+    });
+    return this.userRepository.findOne(new ObjectId(user.id));
   }
 }
